week2/Day3/requirements.txt
pipecat-ai[local,silero,deepgram,openai,elevenlabs,local-smart-turn-v3]
python-dotenv
loguru
